Add files via upload
I was able to set up a postgresql server and set it up to communicate with the app. Using cmd, it would ask the user to input their address and then post it to the server. 
The next step I was trying to figure out was how to then initiate an api search using the zip code as a variable and then return the temperature. Otherwise I was going to try to do a google search and use the Xpath to return the temperature. I was unable to figure out enough information about either of these two options to get it to successfully work.